[Refactor] Domain에서의 Data 의존성 제거
