+++ app/src/main/java/com/hara/kaera/data/datasource/remote/KaeraApi.kt
@@ -2,7 +2,7 @@ package com.hara.kaera.data.datasource.remote
 
 import com.hara.kaera.data.dto.DeleteWorryDTO
 import com.hara.kaera.data.dto.HomeWorryListDTO
-import com.hara.kaera.data.dto.login.KaKaoLoginReqDTO
+import com.hara.kaera.domain.dto.KaKaoLoginReqDTO
 import com.hara.kaera.data.dto.login.KaKaoLoginResDTO
 import com.hara.kaera.data.dto.ReviewReqDTO
 import com.hara.kaera.data.dto.ReviewResDTO
