+++ app/src/main/java/com/hara/kaera/data/datasource/remote/LoginApi.kt
@@ -1,8 +1,8 @@
 package com.hara.kaera.data.datasource.remote
 
-import com.hara.kaera.data.dto.login.JWTRefreshReqDTO
+import com.hara.kaera.domain.dto.JWTRefreshReqDTO
 import com.hara.kaera.data.dto.login.JWTRefreshResDTO
-import com.hara.kaera.data.dto.login.KaKaoLoginReqDTO
+import com.hara.kaera.domain.dto.KaKaoLoginReqDTO
 import com.hara.kaera.data.dto.login.KaKaoLoginResDTO
 import retrofit2.http.Body
 import retrofit2.http.POST
