+++ app/src/main/java/com/hara/kaera/data/datasource/remote/LoginDataSource.kt
@@ -1,8 +1,8 @@
 package com.hara.kaera.data.datasource.remote
 
-import com.hara.kaera.data.dto.login.JWTRefreshReqDTO
+import com.hara.kaera.domain.dto.JWTRefreshReqDTO
 import com.hara.kaera.data.dto.login.JWTRefreshResDTO
-import com.hara.kaera.data.dto.login.KaKaoLoginReqDTO
+import com.hara.kaera.domain.dto.KaKaoLoginReqDTO
 import com.hara.kaera.data.dto.login.KaKaoLoginResDTO
 import kotlinx.coroutines.flow.Flow
