+++ app/src/main/java/com/hara/kaera/data/repository/LoginRepositoryImpl.kt
@@ -8,8 +8,8 @@ import com.hara.kaera.application.Constant
 import com.hara.kaera.application.Constant.EMPTY_TOKEN
 import com.hara.kaera.core.ApiResult
 import com.hara.kaera.data.datasource.remote.LoginDataSource
-import com.hara.kaera.data.dto.login.JWTRefreshReqDTO
-import com.hara.kaera.data.dto.login.KaKaoLoginReqDTO
+import com.hara.kaera.domain.dto.JWTRefreshReqDTO
+import com.hara.kaera.domain.dto.KaKaoLoginReqDTO
 import com.hara.kaera.data.mapper.LoginMapper
 import com.hara.kaera.domain.entity.login.KakaoLoginJWTEntity
 import com.hara.kaera.domain.repository.LoginRepository
